Fixes code review issue
