+++ core-modular/package.json
@@ -8,7 +8,7 @@
     "@eslint/js": "9.39.3",
     "@fast-check/jest": "2.2.0",
     "@happy-dom/jest-environment": "20.8.9",
-    "@luigi-project/container": "^1.7.9",
+    "@luigi-project/container": "^1.7.10",
     "@sveltejs/vite-plugin-svelte": "7.1.2",
     "@tsconfig/svelte": "5.0.8",
     "@types/jest": "30.0.0",
